# docs/03-API/OpenAPI/openapi-finance-v1.yaml
openapi: 3.1.0
info:
  title: Rental Finance API
  version: 1.3.0
  description: 财务域接口（含列表分页/排序/过滤）
servers:
  - url: https://api.rental.example.com
paths:
  /api/v1/invoices:
    get:
      summary: 查询发票列表
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortOrder'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiPageInvoice' }
    post:
      summary: 申请发票
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateInvoiceRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseInvoice' }
  /api/v1/bills:
    get:
      summary: 查询账单列表（B/G端）
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortOrder'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiPageBill' }
    post:
      summary: 生成账单（按账期聚合）
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBillRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseBill' }
  /api/v1/bills/{billId}:
    get:
      summary: 查询账单详情
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - in: path
          name: billId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseBill' }
  /api/v1/bills/{billId}/confirm:
    put:
      summary: 确认账单（进入待支付）
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - in: path
          name: billId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmBillRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponseBill' }
components:
  parameters:
    XRequestId: { in: header, name: X-Request-Id, required: false, schema: { type: string } }
    PageNum: { in: query, name: pageNum, schema: { type: integer, default: 1, minimum: 1 } }
    PageSize: { in: query, name: pageSize, schema: { type: integer, default: 20, minimum: 1, maximum: 200 } }
    SortBy: { in: query, name: sortBy, schema: { type: string, default: createdAt } }
    SortOrder: { in: query, name: sortOrder, schema: { type: string, enum: [asc, desc], default: desc } }
    Filter: { in: query, name: filter, schema: { type: string } }
  schemas:
    ApiBase:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        timestamp: { type: string, format: date-time }
        requestId: { type: string }
      required: [code, message, timestamp, requestId]
    ApiResponseInvoice:
      allOf:
        - $ref: '#/components/schemas/ApiBase'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Invoice' }
    ApiPageInvoice:
      allOf:
        - $ref: '#/components/schemas/ApiBase'
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items: { $ref: '#/components/schemas/Invoice' }
                pageNum: { type: integer }
                pageSize: { type: integer }
                total: { type: integer }
    ApiResponseBill:
      allOf:
        - $ref: '#/components/schemas/ApiBase'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Bill' }
    ApiPageBill:
      allOf:
        - $ref: '#/components/schemas/ApiBase'
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items: { $ref: '#/components/schemas/Bill' }
                pageNum: { type: integer }
                pageSize: { type: integer }
                total: { type: integer }
    CreateInvoiceRequest:
      type: object
      required: [orderId, titleType, invoiceTitle]
      properties:
        orderId: { type: string, format: uuid }
        titleType: { type: string, enum: [PERSONAL, COMPANY] }
        invoiceTitle: { type: string }
        taxNo: { type: string }
        email: { type: string }
        billId: { type: string, format: uuid }
    CreateBillRequest:
      type: object
      required: [billingAccountId, accountType, billingPeriod]
      properties:
        billingAccountId: { type: string, format: uuid }
        accountType: { type: string, enum: [B, G] }
        billingPeriod:
          type: string
          description: 月结账期，格式YYYY-MM
        dueDate: { type: string, format: date }
    ConfirmBillRequest:
      type: object
      required: [confirmedBy]
      properties:
        confirmedBy: { type: string, format: uuid }
        confirmRemark: { type: string }
    Invoice:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        billId: { type: string, format: uuid }
        invoiceNo: { type: string }
        amount: { type: number, format: double }
        status: { type: string }
    Bill:
      type: object
      properties:
        id: { type: string, format: uuid }
        billNo: { type: string }
        billingAccountId: { type: string, format: uuid }
        accountType: { type: string, enum: [B, G] }
        billingPeriod:
          type: string
          description: 月结账期，格式YYYY-MM
        totalAmount: { type: number, format: double }
        paidAmount: { type: number, format: double }
        status: { type: string, enum: [PENDING_CONFIRM, PENDING_PAYMENT, PARTIALLY_PAID, PAID, PAYMENT_FAILED, OVERDUE] }
        confirmedAt: { type: string, format: date-time }
        confirmedBy: { type: string, format: uuid }
        lastPaymentAt: { type: string, format: date-time }
        reconciliationStatus: { type: string, enum: [PENDING, DONE, FAILED] }
        dueDate: { type: string, format: date }